backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
supabase==2.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
groq==0.4.1
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
